requirements.txt
# brew install portaudio
flask
pyyaml
pyaudio
pydub
pygame
openai
ibm-watson
ibm-cloud-sdk-core